React NYT Repository
It is nowhere near finished but this is the start I have made and will update as and when.

Recreating my NYT book review API project by using React.js. Using it as a learning tool to get to grips with hooks such as useEffect and useState, understand more about JSX and also get used to using node.js and packages such as webpack.

N.B. I appreciate this could be more efficient, it is just for teaching myself how to use the different technologies.